+++ Sources/NewmiSharedDTO/PnStatusRs.swift
@@ -5,6 +5,8 @@
 //  Created by Demetris Georgiou on 09/02/2025.
 //
 
+import Foundation
+
 public enum PnStatus: String, Codable, Sendable {
     case registered
     case notRegistered = "not_registered"
@@ -12,9 +14,9 @@ public enum PnStatus: String, Codable, Sendable {
 
 public struct PnStatusRs: Codable, Hashable, Sendable {
     public var status: PnStatus?
-    public var timeline: [Int]?
+    public var timeline: [CGFloat]?
     
-    public init(status: PnStatus? = nil, timeline: [Int]? = nil) {
+    public init(status: PnStatus? = nil, timeline: [CGFloat]? = nil) {
         self.status = status
         self.timeline = timeline
     }
